Fixed issue 18 by calling superClassConstructor1 if superClassConstructor0 doesn't exist.
Also fixed some bugs in PERWAPI that was causing lookup of super_classClassConstructor1 to fail.

git-svn-id: http://rubydotnetcompiler.googlecode.com/svn/trunk@98 f224b4e6-8335-0410-baf7-2d8196006633

+++ src/RubyRuntime/Compiler/CodeGenFacade.cs
@@ -678,7 +678,7 @@ namespace Ruby.Compiler
             else
             {
                 AssemblyName name = type.Assembly.GetName();
-                return cached[type.Assembly] = AssemblyRef.MakeAssemblyRef(name.Name, name.Version.Major, name.Version.Minor,name.Version.Build, name.Version.Revision, name.GetPublicKeyToken());
+                return cached[type.Assembly] = AssemblyRef.MakeAssemblyRef(name.Name, (ushort)name.Version.Major, (ushort)name.Version.Minor, (ushort)name.Version.Build, (ushort)name.Version.Revision, name.GetPublicKeyToken());
             }
         }
